250819 Feat : 리포스트하면 상단에 뜨게 수정

+++ wokkitokki-back/src/main/java/com/winter/wokkitokki/search/document/UserDocument.java
@@ -35,5 +35,5 @@ public class UserDocument {
     private int followingCount;
 
     @Field(type = FieldType.Integer)
-    private int postCount;
+    private Long postCount;
 }
